+++ devops/postgres/upgrade/0.0.5.5.2.sql
@@ -1,0 +1,1 @@
+CREATE  INDEX ON public.node_node_ngrams USING btree (node1_id,  node2_id, ngrams_type);